eni: Refactor limits package
[ upstream commit 6d9cfe48d45bc848f01c278c2054702f60c8fffa ]

This commit refactors the limits package. The biggest change is
consolidating the limits map and its mutex into a single type. Functions
with "limit" in the name have been removed because it is now redundant
since these functions live under the limits package.

The reason why this commit wasn't squashed with the previous is because
git wouldn't consider the new file as a rename from the old, likely
because of too many refactoring changes.

Signed-off-by: Chris Tarazi <chris@isovalent.com>
Signed-off-by: Alexandre Perrin <alex@kaworu.ch>